feat: implement semantic fingerprinting and geometry summary for Flash PDF review
